+++ infra/postgres/migrations/0045_memory_retention_policy.down.sql
@@ -1,0 +1,4 @@
+BEGIN;
+DROP TABLE IF EXISTS memory_retention_policy;
+DELETE FROM schema_migration WHERE version='0045_memory_retention_policy';
+COMMIT;